Added `for_value`, examples, documentation.
