+++ src/raster/tests.rs
@@ -2,8 +2,8 @@ use crate::dataset::Dataset;
 use crate::metadata::Metadata;
 use crate::raster::rasterband::ResampleAlg;
 use crate::raster::{
-    ByteBuffer, ColorEntry, ColorInterpretation, ColorTable, GdalType, GdalTypeDescriptor,
-    RasterCreationOption, StatisticsAll, StatisticsMinMax,
+    ByteBuffer, ColorEntry, ColorInterpretation, ColorTable, RasterCreationOption, StatisticsAll,
+    StatisticsMinMax,
 };
 use crate::test_utils::TempFixture;
 use crate::vsi::unlink_mem_file;
@@ -11,7 +11,6 @@ use crate::DriverManager;
 use gdal_sys::GDALDataType;
 use std::path::Path;
 
-use gdal_sys::GDALDataType::*;
 #[cfg(feature = "ndarray")]
 use ndarray::arr2;
 
@@ -893,76 +892,3 @@ fn test_raster_stats() {
         }
     );
 }
-
-#[test]
-#[allow(non_upper_case_globals)]
-fn test_gdal_data_type() {
-    for t in GdalTypeDescriptor::available_types() {
-        // Test converting from GDALDataType:Type
-        let t2: GdalTypeDescriptor = t.gdal_type().try_into().unwrap();
-        assert_eq!(t, &t2, "{}", t);
-        assert!(t.bits() > 0, "{}", t);
-        assert_eq!(t.bits(), t.bytes() * 8, "{}", t);
-        let name = t.name().unwrap();
-        match t.gdal_type() {
-            GDT_Byte | GDT_UInt16 | GDT_Int16 | GDT_UInt32 | GDT_Int32 => {
-                assert!(t.is_integer(), "{}", &name);
-                assert!(!t.is_floating(), "{}", &name);
-            }
-            GDT_Float32 | GDT_Float64 => {
-                assert!(!t.is_integer(), "{}", &name);
-                assert!(t.is_floating(), "{}", &name);
-            }
-            o => panic!("unknown type ordinal '{}'", o),
-        }
-        match t.gdal_type() {
-            GDT_Byte | GDT_UInt16 | GDT_UInt32 => {
-                assert!(!t.is_signed(), "{}", &name);
-            }
-            GDT_Int16 | GDT_Int32 | GDT_Float32 | GDT_Float64 => {
-                assert!(t.is_signed(), "{}", &name);
-            }
-            o => panic!("unknown type ordinal '{}'", o),
-        }
-    }
-}
-
-#[test]
-fn test_data_type_from_name() {
-    assert!(GdalTypeDescriptor::from_name("foobar").is_err());
-
-    for t in GdalTypeDescriptor::available_types() {
-        let name = t.name().unwrap();
-        let t2 = GdalTypeDescriptor::from_name(&name);
-        assert!(t2.is_ok());
-    }
-}
-
-#[test]
-fn test_data_type_union() {
-    let f32d = <f32>::descriptor();
-    let f64d = <f64>::descriptor();
-
-    let u8d = <u8>::descriptor();
-    let u16d = <u16>::descriptor();
-    let i16d = <i16>::descriptor();
-    let i32d = <i32>::descriptor();
-
-    // reflexivity
-    assert_eq!(i16d.union(i16d), i16d);
-    // symmetry
-    assert_eq!(i16d.union(f32d), f32d);
-    assert_eq!(f32d.union(i16d), f32d);
-    // widening
-    assert_eq!(u8d.union(u16d), u16d);
-    assert_eq!(f32d.union(i32d), f64d);
-
-
-    #[cfg(all(major_ge_3, minor_ge_5))]
-    {
-        let u32d = <u32>::descriptor();
-        let i64d = <i64>::descriptor();
-        assert_eq!(i16d.union(u32d), i64d);
-    }
-}
-
